Add Android Studio overlay.
Recently I've submitted a patch to update the version of Android Studio
for every channel, but until that gets merged, to get access to a newer
version of Android Studio, I've created this overlay.

A bunch of the code had to be copied over from the NixOS Android Studio
code as it uses an internal builder. See a discussion here: https://discourse.nixos.org/t/overlays-for-packages-with-internal-builders/7774